split temp into eval_temp.sh
